requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pymongo==4.6.0
pinecone-client==2.2.4
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.0
openai==1.3.0
whisper==1.1.10
elevenlabs==0.2.26
langchain==0.0.350
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
opencv-python==4.8.1.78
Pillow==10.1.0
nltk==3.8.1
langdetect==1.0.9
google-generativeai==0.3.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

